@@ -48,24 +48,23 @@
     "node": ">=8"
   },
   "devDependencies": {
-    "@angular/animations": "7.2.1",
-    "@angular/common": "7.2.1",
-    "@angular/compiler": "7.2.1",
-    "@angular/compiler-cli": "7.2.1",
-    "@angular/core": "7.2.1",
-    "@angular/forms": "7.2.1",
-    "@angular/http": "7.2.1",
-    "@angular/platform-browser": "7.2.1",
-    "@angular/platform-browser-dynamic": "7.2.1",
-    "@angular/router": "7.2.1",
+    "@angular/animations": "~13.3.11",
+    "@angular/common": "^13.3.11",
+    "@angular/compiler": "^13.3.11",
+    "@angular/compiler-cli": "~13.3.11",
+    "@angular/core": "^13.3.11",
+    "@angular/forms": "^13.3.11",
+    "@angular/platform-browser": "~13.3.11",
+    "@angular/platform-browser-dynamic": "~13.3.11",
+    "@angular/router": "^13.3.11",
     "@babel/core": "7.2.2",
     "@babel/types": "7.3.0",
     "@commitlint/cli": "7.3.2",
     "@commitlint/config-conventional": "7.3.1",
     "@commitlint/prompt-cli": "7.3.1",
-    "@types/jest": "23.3.13",
+    "@types/jest": "27.0.2",
     "babel-jest": "23.6.0",
-    "codelyzer": "4.5.0",
+    "codelyzer": "^5.2.0",
     "docsify-cli": "4.3.0",
     "husky": "1.3.1",
     "jest": "23.6.0",
@@ -74,16 +73,16 @@
     "jest-zone-patch": "0.0.10",
     "lerna": "3.13.4",
     "lint-staged": "8.1.0",
-    "ng-packagr": "4.6.0",
+    "ng-packagr": "^13.3.1",
     "npm-run-all": "4.1.5",
-    "prettier": "1.15.3",
-    "redux": "4.0.1",
+    "prettier": "2.7.1",
+    "redux": "4.0.5",
     "rimraf": "2.6.3",
-    "rxjs": "6.3.3",
+    "rxjs": "~6.6.0",
     "tsickle": "0.34.0",
-    "tslib": "1.9.3",
-    "tslint": "5.12.1",
-    "typescript": "3.2.4",
-    "zone.js": "0.8.28"
+    "tslib": "^2.0.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.6.4",
+    "zone.js": "~0.11.4"
   }
 }
