@@ -3,7 +3,7 @@ import { NgReduxRouter } from './router';
 
 @NgModule()
 export class NgReduxRouterModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<NgReduxRouterModule> {
     return {
       ngModule: NgReduxRouterModule,
       providers: [NgReduxRouter],
